flet_gen_ui/requirements.txt
langchain-openai
langchain-community
langchain
python-dotenv
flet
pyautogen